Cobrand 'hidden_categories' method
This deals with the issue of hiding a two_tier category for the
Hart cobrand, without removing it for the whole of Hampshire,
e.g. see https://github.com/mysociety/FixMyStreet-Commercial/issues/477
but is written as a generic extension to setup_categories_and_bodies
as it may be generally useful.