Explanatory text/link on /my with no reports
Usually you won't get to "Your Reports" page without any
content (reports or updates) but it's possible to do so,
as we just discovered in a client meeting, logging into
that page, and arriving on a screen of blank content with
no guidance.  This patch shows some text and a link to
create a new report, if there aren't any problems/updates.

Fixes #671